refactored data folder and corrected compilation errors
