+++ WatchTvSeries/app/src/main/java/com/example/watchtvseries/data/Repository.kt
@@ -1,0 +1,9 @@
+package com.example.watchtvseries.data
+
+import dagger.hilt.android.scopes.ActivityRetainedScoped
+import javax.inject.Inject
+
+@ActivityRetainedScoped
+class Repository @Inject constructor(remoteData: RemoteData){
+    val remoteDatas = remoteData
+}
