+++ WatchTvSeries/app/src/main/java/com/example/watchtvseries/model/Network.kt
@@ -1,11 +1,11 @@
-package com.example.watchtvseries.Model
+package com.example.watchtvseries.model
 
 
 import com.google.gson.annotations.SerializedName
 
-data class WebChannel(
+data class Network(
     @SerializedName("country")
-    val country: Any,
+    val country: Country,
     @SerializedName("id")
     val id: Int,
     @SerializedName("name")
