+++ WatchTvSeries/app/src/main/java/com/example/watchtvseries/model/Previousepisode.kt
@@ -1,9 +1,9 @@
-package com.example.watchtvseries.Model
+package com.example.watchtvseries.model
 
 
 import com.google.gson.annotations.SerializedName
 
-data class Self(
+data class Previousepisode(
     @SerializedName("href")
     val href: String
 )
